Urls should not be updated during unpublishing phase (#203)
The urls does not need to be updated at all. The reason urls
were presumably updated in the past was to avoid that you could
visit a page if the page was in draft mode.

This has been fixed in django-cms. A page that is in draft
will return an EmptyPageContent, if it is empty it won't render.

See https://github.com/divio/django-cms/pull/6803