Update the change_language_menu to most recent version
The change_language_menu is taken from the most recent version
in django-cms. In essence we don't need to monkeypatch this,
right now we still need this monkeypatch because the work to
overwrite the delete view in pagecontent admin is not yet done.

Once that work is done this monkeypatch is redundant.